chore: fixing grid util functions, modifying srcset

+++ src/data/grid.ts
@@ -1,30 +1,30 @@
 import sassVars from '@/styles/export-vars.module.scss';
 
 export function mobileColumns({ numCols = parseInt(sassVars.gridNumOfColsMobile), extraGutters = 0 }) {
-  const extraGuttersWidth = parseInt(sassVars.gridGapMobile) * extraGutters;
+  const extraGuttersWidth = parseFloat(sassVars.gridGapMobile) * extraGutters;
   return `${
-    (parseInt(sassVars.gridColumnWidthMobile) * numCols +
-      parseInt(sassVars.gridGapMobile) * Math.max(numCols - 1, 0) +
+    (parseFloat(sassVars.gridColumnWidthMobile) * numCols +
+      parseFloat(sassVars.gridGapMobile) * Math.max(numCols - 1, 0) +
       extraGuttersWidth) *
     100
   }vw`;
 }
 
 export function tabletColumns({ numCols = parseInt(sassVars.gridNumOfColsTablet), extraGutters = 0 }) {
-  const extraGuttersWidth = parseInt(sassVars.gridGapTablet) * extraGutters;
+  const extraGuttersWidth = parseFloat(sassVars.gridGapTablet) * extraGutters;
   return `${
-    (parseInt(sassVars.gridColumnWidthTablet) * numCols +
-      parseInt(sassVars.gridGapTablet) * Math.max(numCols - 1, 0) +
+    (parseFloat(sassVars.gridColumnWidthTablet) * numCols +
+      parseFloat(sassVars.gridGapTablet) * Math.max(numCols - 1, 0) +
       extraGuttersWidth) *
     100
   }vw`;
 }
 
 export function desktopColumns({ numCols = parseInt(sassVars.gridNumCols), extraGutters = 0 }) {
-  const extraGuttersWidth = parseInt(sassVars.gridGapDesktop) * extraGutters;
+  const extraGuttersWidth = parseFloat(sassVars.gridGapDesktop) * extraGutters;
   return `${
-    (parseInt(sassVars.gridColumnWidthDesktop) * numCols +
-      parseInt(sassVars.gridGapDesktop) * Math.max(numCols - 1, 0) +
+    (parseFloat(sassVars.gridColumnWidthDesktop) * numCols +
+      parseFloat(sassVars.gridGapDesktop) * Math.max(numCols - 1, 0) +
       extraGuttersWidth) *
     100
   }vw`;
